requirements.txt
boto3
python-dotenv
pytest